ES6ify and start on jsdocing the main InfluxDB class
