+++ lib/pool.js
@@ -20,28 +20,34 @@ const resubmitErrorCodes = [
 ]
 
 /**
+ * @typedef {Object} PoolOptions an options object passed to instantiate
+ *     or configure a pool of Influx connections.
+ * @param {Object} [options.failoverTimeout=60000] the length of time a
+ *     host should be removed for upon a connection error
+ * @param {Number} [options.maxRetries=2] number of times we should retry
+ *     running a query before calling back with an error
+ * @param {Object} [options.requestTimeout=30000] the length of time after
+ *     which HTTP requests will error if they do not receive a response
+ * @param {Function} [options.request] function called to make an HTTP
+ *     request, defaults to the `request` module
+ * @param {Object} [options.backoff] configuration for the backoff strategy
+ *     used when connections fail. It contains a `kind` with other config
+ *     data for the specific kind.
+ * @param {String} [options.backoff.kind] the name of the backoff strategy
+ *     used.
+ */
+
+/**
  *
  * The Pool maintains a list available Influx hosts and dispatches requests
  * to them. If there are errors connecting to hosts, it will disable that
  * host for a period of time.
  */
 class Pool {
+
   /**
    * Creates a new Pool instance.
-   * @param {Object} options
-   * @param {Object} [options.failoverTimeout=60000] the length of time a
-   * host should be removed for upon a connection error
-   * @param {Number} [options.maxRetries=2] number of times we should retry
-   * running a query before calling back with an error
-   * @param {Object} [options.requestTimeout=30000] the length of time after
-   * which HTTP requests will error if they do not receive a response
-   * @param {Function} [options.request] function called to make an HTTP
-   * request, defaults to the `request` module
-   * @param {Object} [options.backoff] configuration for the backoff strategy
-   * used when connections fail. It contains a `kind` with other config
-   * data for the specific kind.
-   * @param {String} [options.backoff.kind] the name of the backoff strategy
-   * used.
+   * @param {PoolOptions} options
    */
   constructor (options) {
     this.options = _.defaults(options, {
@@ -78,7 +84,7 @@ class Pool {
    * @return {Host[]}
    */
   getHostsAvailable () {
-    return this._hostsAvailable
+    return this._hostsAvailable.slice()
   }
 
   /**
@@ -87,7 +93,7 @@ class Pool {
    * @return {Host[]}
    */
   getHostsDisabled () {
-    return this._hostsDisabled
+    return this._hostsDisabled.slice()
   }
 
   /**
@@ -199,9 +205,9 @@ class Pool {
 
   /**
    * Runs a GET request against an Influx server.
-   * @param  {Object}   options  options for the `request` library. Note that
-   * the baseUrl for the host will be set automatically.
-   * @param  {Function} callback
+   * @param {Object} options  options for the `request` library. Note that
+   *     the baseUrl for the host will be set automatically.
+   * @param {Function} callback
    */
   get (options, callback) {
     this._request(options, callback)
@@ -209,9 +215,9 @@ class Pool {
 
   /**
    * Runs a POST request against an Influx server.
-   * @param  {Object}   options  options for the `request` library. Note that
-   * the baseUrl for the host will be set automatically.
-   * @param  {Function} callback
+   * @param {Object} options  options for the `request` library. Note that
+   *      baseUrl for the host will be set automatically.
+   * @param {Function} callback
    */
   post (options, callback) {
     options.method = 'POST'
